Propagate task errors to API client
In the old setup, the task uuid was removed from the Annotation, Variation,
or Coverage object. This had the side effect that exceptions raised in the
task could not longer be catched in the API. We now leave the task uuid there
so we can always look back at the last task result.