Don't store total/variant coverage and coverage threshold
As for coverage values, correct calculation is difficult. Meaning is often
unclear or the information is missing entirely. And in the end I don't see
us really using it anyway. One possible use case would be deducing if we
have a homozygous or heterozygous variant, but I'm doubtfull if we could get
that right at all.

As for the coverage threshold (variant calling minimum coverage depth), we
will probably never use it and variant calling (and filtering) is much too
complex for this simple value to give any insight in.

We really have to leave any analysis based on these things at the origin of
data side, not in Varda.