Refactor variation import
Factor out the Variant model and have everything in the Observation model.
Now we don't have to check if a Variant instance exists and add it if not,
this required us to always commit after every record.
Commit only after every N records, this has quite a big speed increase.