Periodically commit instead of flush in import tasks
Reintroduce the CleanTask base class providing cleanup handles. This is
used in the variation and coverage import tasks, to cleanup when aborting
the import.
Previously, we just used a rollback, but since we now periodically commit
this is not enough. Periodically flushing did not prevent the memory
usage from increasing during task execution, so there's really no choice.
See the variation import task code comments for more information.