Fix race conditions using genome in Celery workers
When `REFERENCE_MISMATCH_ABORT=True`, using the Celery prefork
worker pool with concurrency at least 2, importing many VCF files
at the same time would yield errors like the following:

    TaskError: ('invalid_observations', 'Sequence "G" does not match reference genome on "chrX" at position 74546623')

The reason for this was that file descriptors are inherited after
fork, and thus shared between worker processes. This caused race
conditions with seek and read operations on the genome FASTA file
because it was opened before forking.

The fix is to duplicate the open file object of the FASTA file
after forking.